+++ package.json
@@ -5,7 +5,7 @@
   "main": "js/router.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "devserver": "webpack-dev-server --debug --hot --devtool eval-source-map --output-pathinfo --watch --colors --inline --content-base public --port 8090 --host 10.101.100.184 -s ./src/fonts",
+    "devserver": "webpack-dev-server --debug --hot --devtool eval-source-map --output-pathinfo --watch --colors --inline --content-base public --port 8090 --host 192.168.1.39 -s ./src/fonts",
     "build": "webpack -p"
   },
   "author": "Gorshkov Ivan",
@@ -15,12 +15,13 @@
     "events": "^1.1.1",
     "keymirror": "^0.1.1",
     "material-ui": "^0.15.4",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
-    "react-router": "^2.8.1",
+    "react": "^15.3.1",
+    "react-dom": "^15.3.1",
+    "react-router": "^2.8.0",
     "react-scrollbar": "^0.4.1",
     "react-tap-event-plugin": "^1.0.0",
-    "reflux": "^0.4.1"
+    "reflux": "^0.4.1",
+    "reflux-state-mixin": "^0.7.0"
   },
   "devDependencies": {
     "autoprefixer-loader": "^3.2.0",
@@ -32,12 +33,12 @@
     "css-loader": "^0.25.0",
     "file-loader": "^0.9.0",
     "json-loader": "^0.5.4",
-    "node-sass": "^3.10.0",
-    "react-hot-loader": "^3.0.0-beta.4",
+    "node-sass": "^3.9.3",
+    "react-hot-loader": "^3.0.0-beta.3",
     "sass-loader": "^4.0.2",
     "style-loader": "^0.13.1",
     "url-loader": "^0.5.7",
     "webpack": "^1.13.2",
-    "webpack-dev-server": "^1.16.1"
+    "webpack-dev-server": "^1.15.1"
   }
 }
