--- src/css/views/LogIn.scss
@@ -1,4 +1,0 @@
-.login {
-  display: inline-block; margin: 6px 15px; float: right;
-  .g-input { height: 36px; }
-}
