--- src/db/userData.json
@@ -1,14 +1,0 @@
-[
-  {
-    "id": 1,
-    "avatar": "",
-    "surname": "Горшков",
-    "firstName": "Иван",
-    "lastName": "Андреевич",
-    "nickname": "nextgennexia",
-    "lessonsData": {
-      "newWords": 1,
-      "newWordsPractice": 1
-    }
-  }
-]
