+++ src/js/App.jsx
@@ -1,8 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router';
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import injectTapEventPlugin from 'react-tap-event-plugin';
-import LogIn from './views/LogIn.jsx';
 import '../css/App.scss'
 
 injectTapEventPlugin();
@@ -10,18 +8,12 @@ injectTapEventPlugin();
 export default React.createClass({
   render: function () {
     return (
-      <MuiThemeProvider>
-        <div className='app'>
-          <div className='header'>
-            <Link className='gaku-logo' to='/'>学APP</Link>
-            <LogIn />
-          </div>
-          <div className='navbar'>
-
-          </div>
-          {this.props.children}
+      <div className='app'>
+        <div　className='header'>
+          <Link className='gaku-logo' to='/'>学APP</Link>
         </div>
-      </MuiThemeProvider>
+        {this.props.children}
+      </div>
     );
   }
 });
