+++ src/js/actions/Actions.js
@@ -1,7 +1,5 @@
 import Reflux from 'reflux';
 
 export default Reflux.createActions([
-  'changeDictionary',
-  'chooseLesson',
-  'login'
+  'changeDictionary'
 ]);
