+++ src/js/components/Input.jsx
@@ -1,7 +1,11 @@
 import React from 'react';
+import StateMixin from 'reflux-state-mixin';
+import Actions from '../actions/Actions';
+import DictionaryStores from '../stores/DictionaryStore';
 import '../../css/components/Input.scss';
 
 export default React.createClass({
+  mixins: [StateMixin.connect(DictionaryStores)],
   getDefaultProps: function () {
     return {
       placeholder: ''
@@ -12,8 +16,21 @@ export default React.createClass({
       text: ''
     };
   },
-  onChange: function(event) {
-    this.setState({text: event.target.value});
+  convertToKana: function (event) {
+    let
+      value = event.target.value,
+      dictionary = this.state.dictionary;
+
+    for (let key in dictionary) {
+      value = value.replace(new RegExp(key, 'g'), dictionary[key])
+    }
+
+    this.setState({text: value});
+  },
+  changeDictionary: function (event) {
+    if ((event.keyCode === ' '.charCodeAt(0) && event.ctrlKey) || this.props.changeDictionary) {
+      Actions.changeDictionary();
+    }
   },
   render: function () {
     return (
@@ -22,7 +39,8 @@ export default React.createClass({
           type='text'
           placeholder={this.props.placeholder}
           value={this.state.text}
-          onChange={this.onChange}
+          onChange={this.convertToKana}
+          onKeyDown={this.changeDictionary}
         />
         <div>
           <hr />
