+++ src/js/views/lessons/IntroductionPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ToggleButton from '../../components/ToggleButton.jsx';
-import Input from '../../components/InputTranslator.jsx';
+import Input from '../../components/Input.jsx';
 import RaisedButton from '../../components/RaisedButton.jsx';
 
 export default React.createClass({
